fix: Broken test cases
